Revert "add MusicPlayer basic functionality"
