+++ package.json
@@ -18,14 +18,8 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "@discord-player/extractor": "^7.1.0",
-    "@discordjs/voice": "^0.18.0",
-    "discord-player": "^7.1.0",
-    "discord-player-youtubei": "^1.4.6",
     "discord.js": "^14.19.3",
     "dotenv": "^16.5.0",
-    "openai": "^4.100.0",
-    "play-dl": "^1.9.7",
-    "spotify-web-api-node": "^5.0.2"
+    "openai": "^4.100.0"
   }
 }
