+++ src/discord_tools/command.js
@@ -1,58 +1,10 @@
-const { REST, Routes, Collection, Events, MessageFlags } = require('discord.js');
+const { REST, Routes } = require('discord.js');
 const fsPromises = require('node:fs').promises;
-const fs = require('node:fs');
+
 const path = require('node:path');
 const dotenv = require('dotenv');
 
-const client = require('./client');
-
 dotenv.config();
-
-
-
-client.commands = new Collection();
-
-const foldersPath1 = path.join(__dirname, "commands");
-const commandFolders = fs.readdirSync(foldersPath1);
-
-for (const folder of commandFolders) {
-	const commandsPath = path.join(foldersPath1, folder);
-	const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('.js'));
-	for (const file of commandFiles) {
-		const filePath = path.join(commandsPath, file);
-		const command = require(filePath);
-		// Set a new item in the Collection with the key as the command name and the value as the exported module
-		if ('data' in command && 'execute' in command) {
-			client.commands.set(command.data.name, command);
-		} else {
-			console.log(`[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`);
-		}
-	}
-}
-
-
-client.on(Events.InteractionCreate, async interaction => {
-	if (!interaction.isChatInputCommand()) return;
-
-	const command = interaction.client.commands.get(interaction.commandName);
-
-	if (!command) {
-		console.error(`No command matching ${interaction.commandName} was found.`);
-		return;
-	}
-
-	try {
-		await command.execute(interaction);
-	} catch (error) {
-		console.error(error);
-		if (interaction.replied || interaction.deferred) {
-			await interaction.followUp({ content: 'There was an error while executing this command!', flags: MessageFlags.Ephemeral });
-		} else {
-			await interaction.reply({ content: 'There was an error while executing this command!', flags: MessageFlags.Ephemeral });
-		}
-	}
-});
-
 
 const commands = [];
 // Grab all the command folders from the commands directory you created earlier
